+++ R/dashboardPage.R
@@ -10,6 +10,7 @@
 #' @param title App title.
 #' @param old_school Whether to use the wonderful sketchy design for BS4. FALSE
 #' by default.
+#' @param sidebar_collapsed Whether the sidebar is collapsed of not at start. FALSE by default.
 #'
 #' @examples
 #' if(interactive()){
@@ -34,7 +35,11 @@
 #' @export
 bs4DashPage <- function(navbar = NULL, sidebar = NULL, body = NULL, 
                         controlbar = NULL, footer = NULL, title = NULL,
-                        old_school = FALSE){
+                        old_school = FALSE, sidebar_collapsed = FALSE){
+  
+  if (!is.null(sidebar)) {
+    sidebar_cl <- if (sidebar_collapsed) "sidebar-collapse" else "sidebar-open"
+  }
 
   shiny::tags$html(
     # Head
@@ -46,37 +51,22 @@ bs4DashPage <- function(navbar = NULL, sidebar = NULL, body = NULL,
       ),
       shiny::tags$meta(`http-equiv` = "x-ua-compatible", content = "ie=edge"),
 
-      shiny::tags$title(title),
-
-      # icons
-      shiny::includeCSS(system.file("css/font-awesome.min.css", package = "bs4Dash")),
-      shiny::tags$link(href = "https://use.fontawesome.com/releases/v5.0.13/css/all.css", rel = "stylesheet"),
-      shiny::tags$link(href = "http://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css", rel = "stylesheet"),
-      # AdminLTE3
-      shiny::includeCSS(system.file("css/adminlte.min.css", package = "bs4Dash")),
-      
-      # Old school design (bootswatch sketchy design)
-      if (old_school) {
-        shiny::tags$link(href = "https://bootswatch.com/4/sketchy/bootstrap.min.css", rel = "stylesheet")
-      },
-      shiny::tags$link(href = "https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700", rel = "stylesheet"),
-      # javascript
-      shiny::includeScript(system.file("js/bootstrap.bundle.min.js", package = "bs4Dash")),
-      shiny::includeScript(system.file("js/adminlte.min.js", package = "bs4Dash")),
-      shiny::includeScript(system.file("js/bs4Dash.js", package = "bs4Dash"))
-
+      shiny::tags$title(title)
     ),
     # Body
-    shiny::tags$body(
-      class = "hold-transition sidebar-mini",
-      shiny::tags$div(
-        class = "wrapper",
-        navbar,
-        sidebar,
-        # page content
-        body,
-        controlbar,
-        footer
+    addDeps(
+      theme = old_school,
+      shiny::tags$body(
+        class = paste0("hold-transition sidebar-mini ", sidebar_cl),
+        shiny::tags$div(
+          class = "wrapper",
+          navbar,
+          sidebar,
+          # page content
+          body,
+          controlbar,
+          footer
+        )
       )
     )
   )