+++ R/useful-items.R
@@ -957,6 +957,7 @@ bs4Stars <- function(maxstar = 5, grade, status = "warning") {
 #' @param title Jumbotron title.
 #' @param lead Jumbotron lead.
 #' @param href Jumbrotron external link.
+#' @param btn_name Jumbotron button name.
 #' @param status Jumbotron background color. "primary", "success", "warning", "danger" or "info".
 #'
 #' @author David Granjon, \email{dgranjon@@ymail.com}
@@ -991,7 +992,7 @@ bs4Stars <- function(maxstar = 5, grade, status = "warning") {
 #' }
 #'
 #' @export
-bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL,
+bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL, btn_name = "More",
                           status = c("primary", "warning", "danger", "info", "success")) {
   
   status <- match.arg(status)
@@ -1016,7 +1017,7 @@ bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL,
       href = href,
       target = "_blank",
       role = "button",
-      "More"
+      btn_name
     )
   )
 }
@@ -1195,4 +1196,45 @@ bs4ListGroupItem <- function(..., active = FALSE, disabled = FALSE,
       shiny::tags$small(class = if (isTRUE(active)) NULL else "text-muted", footer)
     )
   }
+}
+
+
+
+
+#' @title BS4 ionicons
+#'
+#' @description Create a ionicon. 
+#'
+#' @param name Name of icon. See \url{https://ionicons.com}.
+#' 
+#' @note Similar to the icon function from shiny.
+#'
+#' @author David Granjon, \email{dgranjon@@ymail.com}
+#'
+#' @examples
+#' if(interactive()){
+#'  library(shiny)
+#'  library(bs4Dash)
+#'  
+#'  shiny::shinyApp(
+#'    ui = bs4DashPage(
+#'      navbar = bs4DashNavbar(),
+#'      sidebar = bs4DashSidebar(),
+#'      controlbar = bs4DashControlbar(),
+#'      footer = bs4DashFooter(),
+#'      title = "test",
+#'      body = bs4DashBody(
+#'       ionicon(name ="heart"),
+#'       ionicon(name ="beer")
+#'     )
+#'    ),
+#'    server = function(input, output) {}
+#'  )
+#' }
+#'
+#' @export
+ionicon <- function(name) {
+  if (is.null(name)) stop("Missing icon name")
+  cl <- paste0("icon ion-md-", name)
+  shiny::tags$i(class = cl)
 }