+++ inst/examples/showcase/old_school/app.R
@@ -1,36 +1,4 @@
----
-title: "bs4Dash Showcase"
-author: "David Granjon"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_document
-vignette: >
-  %\VignetteIndexEntry{bs4Dash Showcase}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-## Dashboard Demo
-
-The demo is hosted [here](http://130.60.24.205/bs4Dash/showcase/classic) or 
-[here](http://130.60.24.205/bs4Dash/showcase/old_school).
-From RStudio, you can also launch the gallery as follows:
-```r
 library(shiny)
-library(shinyWidgets)
-library(bs4Dash)
-library(plotly)
-
-bs4DashGallery()
-```
-
-## Demo code
-
-Below is the code of the showcase. You can pass the argument `old_school = TRUE`
-to set up the old school dashboard theme.
-
-```r
-library(shiny)
-library(fontawesome)
 library(shinyWidgets)
 library(bs4Dash)
 library(plotly)
@@ -45,6 +13,8 @@ y <- rnorm(200)
 
 shiny::shinyApp(
   ui = bs4DashPage(
+    sidebar_collapsed = TRUE,
+    old_school = TRUE,
     navbar = bs4DashNavbar(
       status = "white",
       "I can write text in the navbar!",
@@ -52,7 +22,7 @@ shiny::shinyApp(
         show = TRUE,
         labelText = "!",
         status = "danger",
-        src = "http://www.google.fr",
+        src = "https://www.google.fr",
         bs4DropdownMenuItem(
           text = "message 1",
           date = "today"
@@ -68,7 +38,7 @@ shiny::shinyApp(
       status = "primary",
       title = "bs4Dash",
       brandColor = "primary",
-      url = "http://www.google.fr",
+      url = "https://www.google.fr",
       src = "https://adminlte.io/themes/AdminLTE/dist/img/user2-160x160.jpg",
       elevation = 3,
       opacity = 0.8,
@@ -148,7 +118,7 @@ shiny::shinyApp(
               labelTooltip = "Hi Bro!",
               dropdownIcon = "wrench",
               dropdownMenu = dropdownItemList(
-                dropdownItem(url = "http://www.google.com", name = "Link to google"),
+                dropdownItem(url = "https://www.google.com", name = "Link to google"),
                 dropdownItem(url = "#", name = "item 2"),
                 dropdownDivider(),
                 dropdownItem(url = "#", name = "item 3")
@@ -376,15 +346,15 @@ shiny::shinyApp(
                 width = 12,
                 bs4CarouselItem(
                   active = TRUE,
-                  src = "http://placehold.it/900x500/39CCCC/ffffff&text=I+Love+Bootstrap"
+                  src = "https://placehold.it/900x500/39CCCC/ffffff&text=I+Love+Bootstrap"
                 ),
                 bs4CarouselItem(
                   active = FALSE,
-                  src = "http://placehold.it/900x500/3c8dbc/ffffff&text=I+Love+Bootstrap"
+                  src = "https://placehold.it/900x500/3c8dbc/ffffff&text=I+Love+Bootstrap"
                 ),
                 bs4CarouselItem(
                   active = FALSE,
-                  src = "http://placehold.it/900x500/f39c12/ffffff&text=I+Love+Bootstrap"
+                  src = "https://placehold.it/900x500/f39c12/ffffff&text=I+Love+Bootstrap"
                 )
               )
             )
@@ -479,8 +449,8 @@ shiny::shinyApp(
                   title = "Item 3",
                   icon = "paint-brush",
                   status = "warning",
-                  bs4TimelineItemMedia(src = "http://placehold.it/150x100"),
-                  bs4TimelineItemMedia(src = "http://placehold.it/150x100")
+                  bs4TimelineItemMedia(src = "https://placehold.it/150x100"),
+                  bs4TimelineItemMedia(src = "https://placehold.it/150x100")
                 ),
                 bs4TimelineStart(status = "danger")
               )
@@ -508,8 +478,8 @@ shiny::shinyApp(
                 title = "Item 3",
                 icon = "paint-brush",
                 status = "warning",
-                bs4TimelineItemMedia(src = "http://placehold.it/150x100"),
-                bs4TimelineItemMedia(src = "http://placehold.it/150x100")
+                bs4TimelineItemMedia(src = "https://placehold.it/150x100"),
+                bs4TimelineItemMedia(src = "https://placehold.it/150x100")
               ),
               bs4TimelineStart(status = "danger")
             )
@@ -534,7 +504,7 @@ shiny::shinyApp(
             "It uses utility classes for typography and spacing 
             to space content out within the larger container.",
             status = "primary",
-            href = "http://www.google.fr"
+            href = "https://www.google.fr"
           ),
           
           br(),
@@ -569,21 +539,21 @@ shiny::shinyApp(
                 active = TRUE, 
                 disabled = FALSE, 
                 type = "action",
-                src = "http://www.google.fr"
+                src = "https://www.google.fr"
               ),
               bs4ListGroupItem(
                 active = FALSE, 
                 disabled = FALSE, 
                 type = "action",
                 "Dapibus ac facilisis in",
-                src = "http://www.google.fr"
+                src = "https://www.google.fr"
               ),
               bs4ListGroupItem(
                 "Morbi leo risus",
                 active = FALSE, 
                 disabled = TRUE, 
                 type = "action",
-                src = "http://www.google.fr"
+                src = "https://www.google.fr"
               )
             ),
             bs4ListGroup(
@@ -633,7 +603,10 @@ shiny::shinyApp(
       )
     ),
     footer = bs4DashFooter(
-      copyrights = "@DivadNojnarg",
+      copyrights = a(
+        href = "https://twitter.com/divadnojnarg", 
+        target = "_blank", "@DivadNojnarg"
+      ),
       right_text = "2018"
     ),
     title = "bs4Dash Showcase"
@@ -702,5 +675,5 @@ shiny::shinyApp(
     })
     
   }
+  
 )
-```